+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.6] - 2020-11-23
+### Fixed
+CAIR-154: Race condition may cause node instructions to run multiple times
+
 ## [2.0.5] - 2020-11-17
 ### Added
 Added timeout for REST call to get instructions